--- RHI/Vulkan/Logging.cpp
@@ -1,36 +1,0 @@
-#include "Logging.h"
-
-VKAPI_ATTR VkBool32 VKAPI_CALL
-vkInit::debugCallback(VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity,
-                      VkDebugUtilsMessageTypeFlagsEXT messageType,
-                      const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData,
-                      void* pUserData) {
-    std::cerr << "validation layer: " << pCallbackData->pMessage << std::endl;
-
-    return VK_FALSE;
-}
-
-vk::DebugUtilsMessengerEXT vkInit::make_debug_messenger(
-    vk::Instance& instance, vk::DispatchLoaderDynamic& dldi) {
-
-    /*
-    * DebugUtilsMessengerCreateInfoEXT( VULKAN_HPP_NAMESPACE::DebugUtilsMessengerCreateFlagsEXT flags_           = {},
-                                    VULKAN_HPP_NAMESPACE::DebugUtilsMessageSeverityFlagsEXT messageSeverity_ = {},
-                                    VULKAN_HPP_NAMESPACE::DebugUtilsMessageTypeFlagsEXT     messageType_     = {},
-                                    PFN_vkDebugUtilsMessengerCallbackEXT                    pfnUserCallback_ = {},
-                                    void * pUserData_ = {} )
-    */
-
-    vk::DebugUtilsMessengerCreateInfoEXT createInfo =
-        vk::DebugUtilsMessengerCreateInfoEXT(
-            vk::DebugUtilsMessengerCreateFlagsEXT(),
-            vk::DebugUtilsMessageSeverityFlagBitsEXT::eVerbose |
-                vk::DebugUtilsMessageSeverityFlagBitsEXT::eWarning |
-                vk::DebugUtilsMessageSeverityFlagBitsEXT::eError,
-            vk::DebugUtilsMessageTypeFlagBitsEXT::eGeneral |
-                vk::DebugUtilsMessageTypeFlagBitsEXT::eValidation |
-                vk::DebugUtilsMessageTypeFlagBitsEXT::ePerformance,
-            debugCallback, nullptr);
-
-    return instance.createDebugUtilsMessengerEXT(createInfo, nullptr, dldi);
-}
