+++ RHI/Vulkan/engine.h
@@ -9,29 +9,28 @@ class Engine {
     ~Engine();
 
    private:
-    // whether to print debug messages in functions
+    //whether to print debug messages in functions
     bool debugMode = true;
 
-    // glfw window parameters
+    //glfw-related variables
     int width{640};
     int height{480};
     GLFWwindow* window{nullptr};
 
-    // vulkan instance
+    //instance-related variables
     vk::Instance instance{nullptr};
-
-    // debug callback
     vk::DebugUtilsMessengerEXT debugMessenger{nullptr};
-
-    // dynamic instance dispatcher
     vk::DispatchLoaderDynamic dldi;
 
-    // glfw setup
+    //device-related variables
+    vk::PhysicalDevice physicalDevice{nullptr};
+
+    //glfw setup
     void build_glfw_window();
 
-    // instance setup
+    //instance setup
     void make_instance();
 
-    // debug messenger
-    void make_debug_messenger();
+    //device setup
+    void make_device();
 };
